% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resume.R
\name{bppr_resume}
\alias{bppr_resume}
\title{Resume Sampling of BPPR Parameters}
\usage{
bppr_resume(
  object,
  append = FALSE,
  n_post = 1000,
  n_burn = 9000,
  n_adapt = 0,
  n_thin = 1,
  print_every = 1000
)
}
\arguments{
\item{object}{an object of class \code{"bppr"} containing the values of the current Markov chain, the latest of which constitute initial values for the new chain.}

\item{append}{logical; if \code{TRUE}, new samples will be appended to samples in \code{object}. Otherwise, output will only include new samples.}

\item{n_post}{number of posterior draws to obtain from the Markov chain after burn-in.}

\item{n_burn}{number of draws to burn before obtaining \code{n_post} draws for inference. If \code{prior_coefs == "flat"} then these are absorbed into the adapt phase.}

\item{n_adapt}{number of adaptive MCMC iterations to perform before burn-in. Skips sampling basis coefficients and residual variance to save time.}

\item{n_thin}{keep every n_thin posterior draws after burn-in.}

\item{print_every}{print the iteration number every print_every iterations. Use \code{print_every = 0} to silence.}
}
\value{
An object of class \code{"bppr"}. Predictions can be obtained by passing the entire object to the \code{predict.bppr} function.
}
\description{
Resumes fitting a BayesPPR model where the last Markov chain left off, using the same model parameters.
}
\details{
Continues exploring BayesPPR model space using RJMCMC, beginning at the values contained in the last iteration of \code{object}.
}
\seealso{
\link{predict.bppr} for prediction.
}
\keyword{nonparametric}
\keyword{projection}
\keyword{pursuit}
\keyword{regression}
\keyword{splines}
