% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.bppr}
\alias{predict.bppr}
\title{BayesPPR Predictions}
\usage{
\method{predict}{bppr}(object, newdata, idx_use = NULL, ...)
}
\arguments{
\item{object}{a fitted model, output from the \code{bppr} function.}

\item{newdata}{a matrix of new input values at which to predict.  The columns should correspond to the same variables used in the \code{bppr} function.}

\item{idx_use}{index of Markov samples to use when generating predictions.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A matrix with the same number of rows as \code{newdata} and columns corresponding to all MCMC iterations indexed by idx_use. These are samples from the posterior predictive distribution.
}
\description{
Predict function for BayesPPR. Outputs posterior predictive samples for the desired MCMC iterations.
}
\details{
Bare-bones methods. Could be improved for efficiency.
}
\examples{
# See examples in bppr documentation.

}
\seealso{
\link{bppr} for model fitting.
}
