% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.bppr_pca.R
\name{predict.bppr_pca}
\alias{predict.bppr_pca}
\title{BayesPPR Predictions for MUltivariate Response}
\usage{
\method{predict}{bppr_pca}(object, newdata, idx_use = NULL, n_cores = 1, par_type = "fork", ...)
}
\arguments{
\item{object}{a fitted model, output from the \code{bppr} function.}

\item{newdata}{a matrix of new input values at which to predict.  The columns should correspond to the same variables used in the \code{bppr} function.}

\item{idx_use}{index of Markov samples to use when generating predictions.}

\item{n_cores}{number of cores the user desired to utilize for parallel computation.}

\item{par_type}{a character variable dicatating the type of parallel computation to be performed. Supported values include \code{"fork"}, which uses \code{parallel::mclapply()}, and \code{"socket"}, which uses \code{parallel::parLapply()}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An matrix where the first dimension corresponds to MCMC iterations indexed by idx_use, the second dimensions is the same number of rows as \code{newdata}, and the third corresponds to the dimension of the response.
}
\description{
Predict function for BayesPPR. Outputs the posterior predictive samples for the desired MCMC iterations.
}
\details{
Bare-bones methods. Could be improved for efficiency.
}
\examples{
# See examples in bppr documentation.

}
\seealso{
\link{bppr} for model fitting.
}
