% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.bppr}
\alias{print.bppr}
\title{Print BPPR Details}
\usage{
\method{print}{bppr}(x, ...)
}
\arguments{
\item{x}{a \code{bppr} object, returned from \code{bppr}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
no return value
}
\description{
Print some of the details of a BPPR model.
}
