% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.bppr}
\alias{summary.bppr}
\title{Summarize BPPR Details}
\usage{
\method{summary}{bppr}(object, ...)
}
\arguments{
\item{object}{a \code{bppr} object, returned from \code{bppr}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
no return value
}
\description{
Summarize some of the details of a BPPR model.
}
