% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bppr.R
\name{bppr}
\alias{bppr}
\title{Bayesian Projection Pursuit Regression (BayesPPR)}
\usage{
bppr(
  X,
  y,
  n_ridge_mean = 10,
  n_ridge_max = NULL,
  n_act_max = NULL,
  df_spline = 4,
  prob_relu = 2/3,
  var_coefs_shape = 0.5,
  var_coefs_rate = length(y)/2,
  n_dat_act_min = NULL,
  proj_dir_prop_scale = NULL,
  n_act_w_init = NULL,
  feat_w_init = NULL,
  n_draws = 10000,
  model = "bppr"
)
}
\arguments{
\item{X}{a data frame or matrix of predictors. Categorical features should be coded as numeric.}

\item{y}{a numeric response vector.}

\item{n_ridge_mean}{mean for Poisson prior on the number of ridge functions.}

\item{n_ridge_max}{maximum number of ridge functions allowed in the model. Used to avoid memory overload. Defaults to 150 unless the number of observed responses is small.}

\item{n_act_max}{maximum number of active variables in any given ridge function. Defaults to 3 unless categorical features are detected, in which case the default is larger.}

\item{df_spline}{degrees of freedom for spline basis. Stability should be examined for anything other than 4.}

\item{prob_relu}{prior probability that any given ridge function uses a relu transformation.}

\item{var_coefs_shape}{shape for IG prior on the variance of the basis function coefficients. Default is for the Zellner-Siow prior.}

\item{var_coefs_rate}{rate for IG prior on the variance of the basis function coefficients. Default is for the Zellner-Siow prior.}

\item{n_dat_act_min}{minimum number of observed non-zero values in a ridge function. Defaults to 20 or 0.1 times the number of observations, whichever is smaller.}

\item{proj_dir_prop_scale}{scale parameter for generating proposed projection directions. Should be in (0, 1); default is about 0.002.}

\item{n_act_w_init}{vector of initial weights for number of active variables in a ridge function, used in generating proposed basis functions. Default is \code{rep(1, n_act_max)}.}

\item{feat_w_init}{vector of initial weights for features to be used in generating proposed basis functions. Default is \code{rep(1, ncol(X))}.}

\item{n_draws}{number of draws to obtain from the Markov chain.}

\item{model}{"bppr" is the only valid option as of now.}
}
\value{
An object of class 'bppr'. Predictions can be obtained by passing the entire object to the predict.bppr function.
}
\description{
Fits a BayesPPR model using RJMCMC. Can handle categorical features.
}
\details{
Explores BayesPPR model space using RJMCMC. The BayesPPR model has \deqn{y = f(x) + \epsilon,  ~~\epsilon \sim N(0,\sigma^2)} \deqn{f(x) = \beta_0 + \sum_{j=1}^M \beta_j B_j(x)} and \eqn{B_j(x)} is a natural spline basis expansion. We use priors \deqn{\beta \sim N(0,\sigma^2/\tau (B'B)^{-1})} \deqn{M \sim Poisson(\lambda)} as well as the hyper-prior on the variance \eqn{\tau} of the coefficients \eqn{\beta} mentioned in the arguments above.
}
\examples{
\dontrun{
  ################################
  ### univariate example
  ################################
  ## simulate data (Friedman function)
  f <- function(X){
    10*sin(pi*X[,1]*X[,2]) + 20*(X[,3] - .5)^2 + 10*X[,4] + 5*X[,5]
  }
  n <- 500 # number of observations
  p <- 10 #10 variables, only first 5 matter
  X <- matrix(runif(n*p), n, p)
  y <- f(X) + rnorm(n)

  ## fit BPPR
  fit <- bppr(X, y)

  ## prediction
  X_test <- matrix(runif(n*p), n, p)
  preds <- predict(fit, X_test, n_ignore = 9000) # posterior predictive samples
  true_f <- f(X_test)
  plot(true_f, colMeans(preds), xlab = 'true values', ylab='posterior predictive means')
  abline(a=0, b=1, col=2)
}

## minimal example for CRAN testing
fit <- bppr(matrix(1:10), 1:10, n_draws = 2)
}
\seealso{
\link{predict.bppr} for prediction.
}
\keyword{nonparametric}
\keyword{projection}
\keyword{pursuit}
\keyword{regression}
\keyword{splines}
