% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.bppr}
\alias{plot.bppr}
\title{BPPR Plot Diagnostics}
\usage{
\method{plot}{bppr}(x, quants = c(0.025, 0.975), pred = TRUE, ...)
}
\arguments{
\item{x}{a \code{bppr} object.}

\item{quants}{quantiles for intervals, if desired.  NULL if not desired.}

\item{pred}{logical, whether posterior predictions should be plotted (defaults to TRUE).}

\item{...}{graphical parameters.}
}
\description{
Generate diagnostic plots for BPPR model fit.
}
\details{
The first two plots are trace plots for diagnosing convergence.  The third plot is posterior predicted vs observed, with intervals for predictions.  The fourth plot is a histogram of the residuals (of the posterior mean model), with a red curve showing the assumed Normal density (using posterior mean variance). If \code{pred=FALSE} the third and fourth plots are omitted.
}
\examples{
# See examples in bppr documentation.

}
\seealso{
\link{bppr}, \link{predict.bppr}
}
